- 添加 TabbedPageRenderer 渲染 - 更新 BLEWorker 项目 功能开发 - 更新 Additel.BLE 项目 广播类型修复 - 更新 项目引用

+++ source/Additel.Core/Android/ContextExtensions.cs
@@ -7,9 +7,9 @@ namespace Additel.Core
         public static float ToPixels(this Context context, double value)
         {
             var scale = context.Resources.DisplayMetrics.Density;
-            var converted = value * scale;
+            var converted = (float)(value * scale);
 
-            return (float)converted;
+            return converted;
         }
 
         public static double FromPixels(this Context context, float value)
